+++ packages/amplify-util-uibuilder/event-handlers/handle-PostPull.js
@@ -1,0 +1,14 @@
+const logger = require('../commands/utils/logger');
+
+async function run(context) {
+  try {
+    await context.amplify.invokePluginMethod(context, 'ui-builder', undefined, 'executeAmplifyCommand', [context, 'generateComponents']);
+  } catch (e) {
+    // Swallow all errors from the uibuilder plugin
+    logger.error(e);
+  }
+}
+
+module.exports = {
+  run,
+};