+++ packages/amplify-app/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.2.9](https://github.com/aws-amplify/amplify-cli/compare/amplify-app@4.2.8...amplify-app@4.2.9) (2021-12-17)
+
+**Note:** Version bump only for package amplify-app
+
+
+
+
+
+## [4.2.8](https://github.com/aws-amplify/amplify-cli/compare/amplify-app@4.2.7...amplify-app@4.2.8) (2021-12-02)
+
+**Note:** Version bump only for package amplify-app
+
+
+
+
+
+## [4.2.7](https://github.com/aws-amplify/amplify-cli/compare/amplify-app@4.2.6...amplify-app@4.2.7) (2021-12-01)
+
+**Note:** Version bump only for package amplify-app
+
+
+
+
+
 ## [4.2.6](https://github.com/aws-amplify/amplify-cli/compare/amplify-app@4.2.5...amplify-app@4.2.6) (2021-11-26)
 
 **Note:** Version bump only for package amplify-app