+++ packages/amplify-e2e-tests/CHANGELOG.md
@@ -3,6 +3,72 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.5.7](https://github.com/aws-amplify/amplify-cli/compare/amplify-e2e-tests@3.5.6...amplify-e2e-tests@3.5.7) (2021-12-17)
+
+
+### Bug Fixes
+
+* add missing salt to deployment logicalId ([#9234](https://github.com/aws-amplify/amplify-cli/issues/9234)) ([d4109e2](https://github.com/aws-amplify/amplify-cli/commit/d4109e2e8a18e2e9b787a97b1508efb1bf97fac9))
+* update gql v2 custom transformer loading logic ([#9252](https://github.com/aws-amplify/amplify-cli/issues/9252)) ([f728b4b](https://github.com/aws-amplify/amplify-cli/commit/f728b4bb835674afd32dab7243dd3d826601d333))
+
+
+
+
+
+## [3.5.6](https://github.com/aws-amplify/amplify-cli/compare/amplify-e2e-tests@3.5.5...amplify-e2e-tests@3.5.6) (2021-12-09)
+
+
+
+## 7.6.3 (2021-12-07)
+
+
+### Bug Fixes
+
+* **amplify-provider-awscloudformation:** use in memory meta for export ([#9208](https://github.com/aws-amplify/amplify-cli/issues/9208)) ([b3ca83b](https://github.com/aws-amplify/amplify-cli/commit/b3ca83b9bec986f0fd525d46738b277eb93e4384))
+
+
+
+
+
+## [3.5.5](https://github.com/aws-amplify/amplify-cli/compare/amplify-e2e-tests@3.5.4...amplify-e2e-tests@3.5.5) (2021-12-03)
+
+**Note:** Version bump only for package amplify-e2e-tests
+
+
+
+
+
+## [3.5.4](https://github.com/aws-amplify/amplify-cli/compare/amplify-e2e-tests@3.5.3...amplify-e2e-tests@3.5.4) (2021-12-02)
+
+**Note:** Version bump only for package amplify-e2e-tests
+
+
+
+
+
+## [3.5.3](https://github.com/aws-amplify/amplify-cli/compare/amplify-e2e-tests@3.5.2...amplify-e2e-tests@3.5.3) (2021-12-02)
+
+**Note:** Version bump only for package amplify-e2e-tests
+
+
+
+
+
+## [3.5.2](https://github.com/aws-amplify/amplify-cli/compare/amplify-e2e-tests@3.5.1...amplify-e2e-tests@3.5.2) (2021-12-01)
+
+
+
+## 7.5.5 (2021-11-30)
+
+
+### Bug Fixes
+
+* fixes import category parameters file path ([#9140](https://github.com/aws-amplify/amplify-cli/issues/9140)) ([6a7893e](https://github.com/aws-amplify/amplify-cli/commit/6a7893eb4d8932f8106cef2a311c2c945c5b8195))
+
+
+
+
+
 ## [3.5.1](https://github.com/aws-amplify/amplify-cli/compare/amplify-e2e-tests@3.5.0...amplify-e2e-tests@3.5.1) (2021-11-26)