+++ packages/amplify-frontend-ios/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.3.8](https://github.com/aws-amplify/amplify-cli/compare/amplify-frontend-ios@3.3.7...amplify-frontend-ios@3.3.8) (2021-12-17)
+
+**Note:** Version bump only for package amplify-frontend-ios
+
+
+
+
+
+## [3.3.7](https://github.com/aws-amplify/amplify-cli/compare/amplify-frontend-ios@3.3.6...amplify-frontend-ios@3.3.7) (2021-12-02)
+
+**Note:** Version bump only for package amplify-frontend-ios
+
+
+
+
+
+## [3.3.6](https://github.com/aws-amplify/amplify-cli/compare/amplify-frontend-ios@3.3.5...amplify-frontend-ios@3.3.6) (2021-12-01)
+
+**Note:** Version bump only for package amplify-frontend-ios
+
+
+
+
+
 ## [3.3.5](https://github.com/aws-amplify/amplify-cli/compare/amplify-frontend-ios@3.3.4...amplify-frontend-ios@3.3.5) (2021-11-26)
 
 **Note:** Version bump only for package amplify-frontend-ios