+++ packages/amplify-graphql-migration-tests/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.2.14](https://github.com/aws-amplify/amplify-cli/compare/amplify-graphql-migration-tests@2.2.13...amplify-graphql-migration-tests@2.2.14) (2021-12-17)
+
+**Note:** Version bump only for package amplify-graphql-migration-tests
+
+
+
+
+
+## [2.2.13](https://github.com/aws-amplify/amplify-cli/compare/amplify-graphql-migration-tests@2.2.12...amplify-graphql-migration-tests@2.2.13) (2021-12-03)
+
+**Note:** Version bump only for package amplify-graphql-migration-tests
+
+
+
+
+
+## [2.2.12](https://github.com/aws-amplify/amplify-cli/compare/amplify-graphql-migration-tests@2.2.11...amplify-graphql-migration-tests@2.2.12) (2021-12-02)
+
+**Note:** Version bump only for package amplify-graphql-migration-tests
+
+
+
+
+
+## [2.2.11](https://github.com/aws-amplify/amplify-cli/compare/amplify-graphql-migration-tests@2.2.10...amplify-graphql-migration-tests@2.2.11) (2021-12-01)
+
+**Note:** Version bump only for package amplify-graphql-migration-tests
+
+
+
+
+
 ## [2.2.10](https://github.com/aws-amplify/amplify-cli/compare/amplify-graphql-migration-tests@2.2.9...amplify-graphql-migration-tests@2.2.10) (2021-11-29)
 
 **Note:** Version bump only for package amplify-graphql-migration-tests