+++ packages/amplify-util-mock/CHANGELOG.md
@@ -3,6 +3,62 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.2.22](https://github.com/aws-amplify/amplify-cli/compare/amplify-util-mock@4.2.21...amplify-util-mock@4.2.22) (2021-12-17)
+
+**Note:** Version bump only for package amplify-util-mock
+
+
+
+
+
+## [4.2.21](https://github.com/aws-amplify/amplify-cli/compare/amplify-util-mock@4.2.20...amplify-util-mock@4.2.21) (2021-12-09)
+
+**Note:** Version bump only for package amplify-util-mock
+
+
+
+
+
+## [4.2.20](https://github.com/aws-amplify/amplify-cli/compare/amplify-util-mock@4.2.19...amplify-util-mock@4.2.20) (2021-12-03)
+
+**Note:** Version bump only for package amplify-util-mock
+
+
+
+
+
+## [4.2.19](https://github.com/aws-amplify/amplify-cli/compare/amplify-util-mock@4.2.18...amplify-util-mock@4.2.19) (2021-12-02)
+
+**Note:** Version bump only for package amplify-util-mock
+
+
+
+
+
+## [4.2.18](https://github.com/aws-amplify/amplify-cli/compare/amplify-util-mock@4.2.17...amplify-util-mock@4.2.18) (2021-12-02)
+
+**Note:** Version bump only for package amplify-util-mock
+
+
+
+
+
+## [4.2.17](https://github.com/aws-amplify/amplify-cli/compare/amplify-util-mock@4.2.16...amplify-util-mock@4.2.17) (2021-12-02)
+
+**Note:** Version bump only for package amplify-util-mock
+
+
+
+
+
+## [4.2.16](https://github.com/aws-amplify/amplify-cli/compare/amplify-util-mock@4.2.15...amplify-util-mock@4.2.16) (2021-12-01)
+
+**Note:** Version bump only for package amplify-util-mock
+
+
+
+
+
 ## [4.2.15](https://github.com/aws-amplify/amplify-cli/compare/amplify-util-mock@4.2.14...amplify-util-mock@4.2.15) (2021-11-29)
 
 **Note:** Version bump only for package amplify-util-mock