+++ packages/graphql-elasticsearch-transformer/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [5.2.10](https://github.com/aws-amplify/amplify-cli/compare/graphql-elasticsearch-transformer@5.2.9...graphql-elasticsearch-transformer@5.2.10) (2021-12-17)
+
+**Note:** Version bump only for package graphql-elasticsearch-transformer
+
+
+
+
+
+## [5.2.9](https://github.com/aws-amplify/amplify-cli/compare/graphql-elasticsearch-transformer@5.2.8...graphql-elasticsearch-transformer@5.2.9) (2021-12-03)
+
+**Note:** Version bump only for package graphql-elasticsearch-transformer
+
+
+
+
+
+## [5.2.8](https://github.com/aws-amplify/amplify-cli/compare/graphql-elasticsearch-transformer@5.2.7...graphql-elasticsearch-transformer@5.2.8) (2021-12-02)
+
+**Note:** Version bump only for package graphql-elasticsearch-transformer
+
+
+
+
+
+## [5.2.7](https://github.com/aws-amplify/amplify-cli/compare/graphql-elasticsearch-transformer@5.2.6...graphql-elasticsearch-transformer@5.2.7) (2021-12-01)
+
+**Note:** Version bump only for package graphql-elasticsearch-transformer
+
+
+
+
+
 ## [5.2.6](https://github.com/aws-amplify/amplify-cli/compare/graphql-elasticsearch-transformer@5.2.5...graphql-elasticsearch-transformer@5.2.6) (2021-11-26)
 
 **Note:** Version bump only for package graphql-elasticsearch-transformer